+++ Makefile.pseudomodules
@@ -1,1 +1,2 @@
 PSEUDOMODULES += defaulttransceiver
+PSEUDOMODULES += transport_layer
