+++ examples/rpl_udp/Makefile
@@ -54,11 +54,10 @@ BOARD_BLACKLIST := arduino-due mbed_lpc1768 msb-430 pttu udoo qemu-i386 stm32f0d
 
 USEMODULE += shell
 USEMODULE += shell_commands
-USEMODULE += posix
 USEMODULE += ps
 USEMODULE += vtimer
 USEMODULE += defaulttransceiver
 USEMODULE += rpl
-USEMODULE += destiny
+USEMODULE += udp
 
 include $(RIOTBASE)/Makefile.include
