+++ examples/rpl_udp/udp.c
@@ -26,7 +26,7 @@
 
 #include "thread.h"
 
-#include "destiny/socket.h"
+#include "socket_base/socket.h"
 
 #include "net_help.h"
 
@@ -63,7 +63,7 @@ static void *init_udp_server(void *arg)
     char buffer_main[UDP_BUFFER_SIZE];
     int32_t recsize;
     uint32_t fromlen;
-    int sock = destiny_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
+    int sock = socket_base_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
 
     memset(&sa, 0, sizeof(sa));
 
@@ -72,14 +72,14 @@ static void *init_udp_server(void *arg)
 
     fromlen = sizeof(sa);
 
-    if (-1 == destiny_socket_bind(sock, &sa, sizeof(sa))) {
+    if (-1 == socket_base_bind(sock, &sa, sizeof(sa))) {
         printf("Error bind failed!\n");
-        destiny_socket_close(sock);
+        socket_base_close(sock);
+        return NULL;
     }
 
     while (1) {
-        recsize = destiny_socket_recvfrom(sock, (void *)buffer_main, UDP_BUFFER_SIZE, 0,
-                                          &sa, &fromlen);
+        recsize = socket_base_recvfrom(sock, (void *)buffer_main, UDP_BUFFER_SIZE, 0, &sa, &fromlen);
 
         if (recsize < 0) {
             printf("ERROR: recsize < 0!\n");
@@ -88,7 +88,7 @@ static void *init_udp_server(void *arg)
         printf("UDP packet received, payload: %s\n", buffer_main);
     }
 
-    destiny_socket_close(sock);
+    socket_base_close(sock);
 
     return NULL;
 }
@@ -113,7 +113,7 @@ void udp_send(int argc, char **argv)
     strncpy(text, argv[2], sizeof(text));
     text[sizeof(text) - 1] = 0;
 
-    sock = destiny_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
+    sock = socket_base_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
 
     if (-1 == sock) {
         printf("Error Creating Socket!");
@@ -133,7 +133,7 @@ void udp_send(int argc, char **argv)
     memcpy(&sa.sin6_addr, &ipaddr, 16);
     sa.sin6_port = HTONS(SERVER_PORT);
 
-    bytes_sent = destiny_socket_sendto(sock, (char *)text,
+    bytes_sent = socket_base_sendto(sock, (char *)text,
                                        strlen(text) + 1, 0, &sa,
                                        sizeof(sa));
 
@@ -146,5 +146,5 @@ void udp_send(int argc, char **argv)
                                             &ipaddr));
     }
 
-    destiny_socket_close(sock);
+    socket_base_close(sock);
 }
