+++ sys/Makefile
@@ -46,8 +46,18 @@ endif
 ifneq (,$(filter net_if,$(USEMODULE)))
     DIRS += net/link_layer/net_if
 endif
-ifneq (,$(filter destiny,$(USEMODULE)))
-    DIRS += net/transport_layer/destiny
+ifneq (,$(filter transport_layer,$(USEMODULE)))
+	USEMODULE += udp
+	USEMODULE += tcp
+endif
+ifneq (,$(filter socket_base,$(USEMODULE)))
+    DIRS += net/transport_layer/socket_base
+endif
+ifneq (,$(filter udp,$(USEMODULE)))
+    DIRS += net/transport_layer/udp
+endif
+ifneq (,$(filter tcp,$(USEMODULE)))
+    DIRS += net/transport_layer/tcp
 endif
 ifneq (,$(filter net_help,$(USEMODULE)))
     DIRS += net/crosslayer/net_help
