+++ sys/net/network_layer/sixlowpan/lowpan.c
@@ -43,7 +43,7 @@
 #include "icmp.h"
 
 #include "ieee802154_frame.h"
-#include "destiny/in.h"
+#include "socket_base/in.h"
 #include "net_help.h"
 
 #define ENABLE_DEBUG    (0)
