+++ sys/net/transport_layer/udp/Makefile
@@ -1,0 +1,2 @@
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/socket_base
+include $(RIOTBASE)/Makefile.base
